fix: DI injection
