+++ Directory.Packages.props
@@ -67,7 +67,7 @@
     <!-- Miscellaneous -->
     <PackageVersion Include="Scrutor" Version="6.1.0" />
     <PackageVersion Include="Saunter" Version="0.13.0" />
-    <PackageVersion Include="MediatR" Version="12.5.0" />
+    <PackageVersion Include="MediatR" Version="[12.5.0]" />
     <PackageVersion Include="HtmlSanitizer" Version="9.0.886" />
     <PackageVersion Include="Marten.AspNetCore" Version="8.3.1" />
     <PackageVersion Include="Quartz.AspNetCore" Version="3.14.0" />
