+++ src/Services/Catalog/BookWorm.Catalog.UnitTests/Features/Books/Update/UpdateBookCommandTests.cs
@@ -1,89 +1,12 @@
 ﻿using BookWorm.Catalog.Domain.AggregatesModel.BookAggregate;
 using BookWorm.Catalog.Features.Books.Update;
-using BookWorm.Catalog.Infrastructure.Blob;
 using BookWorm.Catalog.UnitTests.Fakers;
 using BookWorm.Chassis.Exceptions;
-using MediatR;
-using Microsoft.AspNetCore.Http;
 
 namespace BookWorm.Catalog.UnitTests.Features.Books.Update;
 
 public sealed class UpdateBookCommandTests
 {
-    #region PreUpdateBookHandler Tests
-
-    [Test]
-    public async Task GivenFormFileExists_WhenPreProcessing_ThenShouldUploadFileAndSetImageName()
-    {
-        // Arrange
-        var formFile = new Mock<IFormFile>();
-        var blobService = new Mock<IBlobService>();
-        const string expectedImageUrl = "image-url.jpg";
-
-        blobService
-            .Setup(x => x.UploadFileAsync(formFile.Object, It.IsAny<CancellationToken>()))
-            .ReturnsAsync(expectedImageUrl);
-
-        var command = new UpdateBookCommand(
-            Guid.CreateVersion7(),
-            "Test Book",
-            "Description",
-            formFile.Object,
-            10.99m,
-            null,
-            Guid.CreateVersion7(),
-            Guid.CreateVersion7(),
-            [Guid.CreateVersion7()]
-        );
-
-        var handler = new PreUpdateBookHandler(blobService.Object);
-
-        // Act
-        await handler.Process(command, CancellationToken.None);
-
-        // Assert
-        command.ImageName.ShouldBe(expectedImageUrl);
-        blobService.Verify(
-            x => x.UploadFileAsync(formFile.Object, It.IsAny<CancellationToken>()),
-            Times.Once
-        );
-    }
-
-    [Test]
-    public async Task GivenNoFormFile_WhenPreProcessing_ThenShouldNotCallBlobService()
-    {
-        // Arrange
-        var blobService = new Mock<IBlobService>();
-
-        var command = new UpdateBookCommand(
-            Guid.CreateVersion7(),
-            "Test Book",
-            "Description",
-            null,
-            10.99m,
-            null,
-            Guid.CreateVersion7(),
-            Guid.CreateVersion7(),
-            [Guid.CreateVersion7()]
-        );
-
-        var handler = new PreUpdateBookHandler(blobService.Object);
-
-        // Act
-        await handler.Process(command, CancellationToken.None);
-
-        // Assert
-        command.ImageName.ShouldBeNull();
-        blobService.Verify(
-            x => x.UploadFileAsync(It.IsAny<IFormFile>(), It.IsAny<CancellationToken>()),
-            Times.Never
-        );
-    }
-
-    #endregion
-
-    #region UpdateBookHandler Tests
-
     [Test]
     public async Task GivenValidCommand_WhenHandling_ThenShouldUpdateBook()
     {
@@ -308,110 +231,4 @@ public sealed class UpdateBookCommandTests
         // Assert
         book.Image.ShouldBe("existing-image.jpg");
     }
-
-    #endregion
-
-    #region PostUpdateBookHandler Tests
-
-    [Test]
-    public async Task GivenIsRemoveImageTrueAndImageNameExists_WhenProcessing_ThenShouldDeleteFile()
-    {
-        // Arrange
-        var blobService = new Mock<IBlobService>();
-        const string imageName = "image-to-delete.jpg";
-
-        var command = new UpdateBookCommand(
-            Guid.CreateVersion7(),
-            "Test Book",
-            "Description",
-            null,
-            10.99m,
-            null,
-            Guid.CreateVersion7(),
-            Guid.CreateVersion7(),
-            [Guid.CreateVersion7()],
-            true // IsRemoveImage = true
-        )
-        {
-            ImageName = imageName,
-        };
-
-        var handler = new PostUpdateBookHandler(blobService.Object);
-
-        // Act
-        await handler.Process(command, Unit.Value, CancellationToken.None);
-
-        // Assert
-        blobService.Verify(
-            x => x.DeleteFileAsync(imageName, It.IsAny<CancellationToken>()),
-            Times.Once
-        );
-    }
-
-    [Test]
-    public async Task GivenIsRemoveImageFalse_WhenProcessing_ThenShouldNotCallBlobService()
-    {
-        // Arrange
-        var blobService = new Mock<IBlobService>();
-        const string imageName = "image.jpg";
-
-        var command = new UpdateBookCommand(
-            Guid.CreateVersion7(),
-            "Test Book",
-            "Description",
-            null,
-            10.99m,
-            null,
-            Guid.CreateVersion7(),
-            Guid.CreateVersion7(),
-            [Guid.CreateVersion7()] // IsRemoveImage = false
-        )
-        {
-            ImageName = imageName,
-        };
-
-        var handler = new PostUpdateBookHandler(blobService.Object);
-
-        // Act
-        await handler.Process(command, Unit.Value, CancellationToken.None);
-
-        // Assert
-        blobService.Verify(
-            x => x.DeleteFileAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()),
-            Times.Never
-        );
-    }
-
-    [Test]
-    public async Task GivenIsRemoveImageTrueButNoImageName_WhenProcessing_ThenShouldNotCallBlobService()
-    {
-        // Arrange
-        var blobService = new Mock<IBlobService>();
-
-        var command = new UpdateBookCommand(
-            Guid.CreateVersion7(),
-            "Test Book",
-            "Description",
-            null,
-            10.99m,
-            null,
-            Guid.CreateVersion7(),
-            Guid.CreateVersion7(),
-            [Guid.CreateVersion7()],
-            true // IsRemoveImage = true
-        );
-
-        var handler = new PostUpdateBookHandler(blobService.Object);
-
-        // Act
-        await handler.Process(command, Unit.Value, CancellationToken.None);
-
-        // Assert
-        blobService.Verify(
-            x => x.DeleteFileAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()),
-            Times.Never
-        );
-    }
-
-    #endregion
 }
