+++ src/Services/Catalog/BookWorm.Catalog/Features/Books/Create/CreateBookCommand.PreProcessor.cs
@@ -1,0 +1,16 @@
+﻿using MediatR.Pipeline;
+
+namespace BookWorm.Catalog.Features.Books.Create;
+
+public sealed class CreateBookPreProcessor(IBlobService blobService)
+    : IRequestPreProcessor<CreateBookCommand>
+{
+    public async Task Process(CreateBookCommand request, CancellationToken cancellationToken)
+    {
+        if (request.Image is not null)
+        {
+            var imageUrl = await blobService.UploadFileAsync(request.Image, cancellationToken);
+            request.ImageName = imageUrl;
+        }
+    }
+}
