+++ src/Services/Catalog/BookWorm.Catalog/Features/Books/Update/UpdateBookCommand.PreProcessor.cs
@@ -1,0 +1,16 @@
+﻿using MediatR.Pipeline;
+
+namespace BookWorm.Catalog.Features.Books.Update;
+
+public sealed class UpdateBookPreProcessor(IBlobService blobService)
+    : IRequestPreProcessor<UpdateBookCommand>
+{
+    public async Task Process(UpdateBookCommand request, CancellationToken cancellationToken)
+    {
+        if (request.Image is not null)
+        {
+            var imageUrl = await blobService.UploadFileAsync(request.Image, cancellationToken);
+            request.ImageName = imageUrl;
+        }
+    }
+}
